[test] split remaining quantization tests out of test_jit
as title, split remaining quantization tests out of test_jit to reduce
the size of test_jit

[ghstack-poisoned]
